+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - `getAll` extensions method for fetching all changelog items 
 - `groups` extensions property for defining the groups created with the Unreleased section
+- `ktlint` integration
 
 ## [0.3.3]
 ### Added